Consolidates to a single stash picker Adds inline buttons to stash & commit picker Ensures stash view & pickers show proper empty state
