+++ src/commands/diffWithRevisionFrom.ts
@@ -6,8 +6,8 @@ import { GitUri } from '../git/gitUri';
 import { isBranchReference } from '../git/utils/reference.utils';
 import { shortenRevision } from '../git/utils/revision.utils';
 import { showNoRepositoryWarningMessage } from '../messages';
-import { showStashPicker } from '../quickpicks/commitPicker';
 import { showReferencePicker } from '../quickpicks/referencePicker';
+import { showStashPicker } from '../quickpicks/stashPicker';
 import { command, executeCommand } from '../system/-webview/command';
 import { selectionToDiffRange } from '../system/-webview/vscode/editors';
 import { basename } from '../system/path';
