+++ src/commands/openFileAtRevisionFrom.ts
@@ -6,8 +6,8 @@ import { openFileAtRevision } from '../git/actions/commit';
 import { GitUri } from '../git/gitUri';
 import type { GitReference } from '../git/models/reference';
 import { showNoRepositoryWarningMessage } from '../messages';
-import { showStashPicker } from '../quickpicks/commitPicker';
 import { showReferencePicker } from '../quickpicks/referencePicker';
+import { showStashPicker } from '../quickpicks/stashPicker';
 import { command } from '../system/-webview/command';
 import { pad } from '../system/string';
 import { ActiveEditorCommand } from './commandBase';
